MediaIndex/requirements.txt
click-config-file
isort
pycodestyle
pydocstyle

cached_property
xxhash

exifyay
git+http://github.com/smarnach/pyexiftool.git#egg=pyexiftool
git+http://github.com/jed-frey/get_files.git#egg=get_files

versioneer
rq
black
redis

# Testing
pytest
cython
Pillow

# Development 
jupyter
notebook

## Host Requirements
#16.04 18.04: python3
#16.04 18.04: python3-venv
#16.04 18.04: python3-dev

git+https://github.com/jed-frey/pydarknet2.git#egg=pydarknet2
